+++ tests/test_examples.py
@@ -4,6 +4,7 @@ import runpy
 from pathlib import Path
 
 import pytest
+from baybe.surrogate import _ONNX_INSTALLED
 from baybe.utils.chemistry import _MORDRED_INSTALLED, _RDKIT_INSTALLED
 
 _CHEM_INSTALLED = _MORDRED_INSTALLED and _RDKIT_INSTALLED
@@ -14,7 +15,7 @@ example_list = list(Path(".").glob("examples/*/*.py"))
 EXAMPLES = [str(file) for file in example_list]
 
 
-if _CHEM_INSTALLED:
+if _CHEM_INSTALLED and _ONNX_INSTALLED:
 
     @pytest.mark.slow
     @pytest.mark.parametrize("example", EXAMPLES)